Report Generator/requirements.txt
neo4j==4.4.1
psycopg2==2.9.3
pytz==2021.3
PyYAML==6.0
